[test] Car 테스트 코드 클래스 합치기
